services/data-scraper-service/pyproject.toml
[project]
name = "data-scrapper-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.121.0",
    "requests>=2.32.5",
    "httpx>=0.28.1",
    "uvicorn>=0.38.0",
    "python-dotenv>=1.2.1",
]
